serve PortletMethods.js from vaadin-portlet-static.war
