+++ vaadin-portlet/src/main/resources/META-INF/resources/scripts/PortletMethods.js
